# docs/swagger.yaml
basePath: /
definitions:
...
  internal_interfaces_http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_interfaces_http.GenericResponse:
    properties:
      data:
        additionalProperties: true
        type: object
    type: object
...
info:
  contact: {}
  description: Internal deployment platform API.
  title: GOS Release API
  version: "1.0"
paths:
...
  /argocd/applications:
    get:
      description: 查询Applications列表，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 查询Applications列表
      tags:
      - argocd
  /argocd/applications/{id}:
    get:
      description: 获取Application By ID详情，并按统一响应结构返回处理结果。
      parameters:
      - description: 资源 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 获取Application By ID详情
      tags:
      - argocd
  /argocd/applications/{id}/original-link:
    get:
      description: 获取Original Link详情，并按统一响应结构返回处理结果。
      parameters:
      - description: 资源 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 获取Original Link详情
      tags:
      - argocd
  /argocd/applications/sync:
    post:
      consumes:
      - application/json
      description: 同步Applications，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 同步Applications
      tags:
      - argocd
  /argocd/env-bindings:
    get:
      description: 查询Env Bindings列表，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 查询Env Bindings列表
      tags:
      - argocd
    put:
      consumes:
      - application/json
      description: 更新Env Bindings，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 更新Env Bindings
      tags:
      - argocd
  /argocd/instances:
    get:
      description: 查询Instances列表，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 查询Instances列表
      tags:
      - argocd
    post:
      consumes:
      - application/json
      description: 创建Instance，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 创建Instance
      tags:
      - argocd
  /argocd/instances/{id}:
    put:
      consumes:
      - application/json
      description: 更新Instance，并按统一响应结构返回处理结果。
      parameters:
      - description: 资源 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 更新Instance
      tags:
      - argocd
...
schemes:
- http
- https
swagger: "2.0"